+++ amanatApi/Models/ACCSANAD.cs
@@ -16,17 +16,17 @@ namespace amanatApi.Models
         public int NO { get; set; }
         public string TITLE { get; set; }
         public string SANADTYPE { get; set; }
-        public long SANADDATE { get; set; }
+        public int SANADDATE { get; set; }
         public int SANADCONDITION { get; set; }
         public int TOTAL { get; set; }
         public string BASEDON { get; set; }
-        public long CREATEDATE { get; set; }
+        public int CREATEDATE { get; set; }
         public string CREATEUSER { get; set; }
-        public long UPDATEDATE { get; set; }
+        public int UPDATEDATE { get; set; }
         public string UPDATEUSER { get; set; }
         public int SENT { get; set; }
         public string APPROVEDUSER { get; set; }
-        public long APPROVEDDATE { get; set; }
+        public int APPROVEDDATE { get; set; }
         public int YEAR { get; set; }
         public int ORGANIZATION { get; set; }
         public int LOCATION { get; set; }
