+++ amanatApi/Models/DATALOG.cs
@@ -14,7 +14,7 @@ namespace amanatApi.Models
 
         public int ID { get; set; }
         public int ENTITYID { get; set; }
-        public long DATE_CREATE { get; set; }
+        public int DATE_CREATE { get; set; }
         public int USERID { get; set; }
         public int ENTITYSTATE { get; set; }
         public string ENTITYNAME { get; set; }
