+++ amanatApi/Models/ERP_ACCSANAD.cs
@@ -15,14 +15,14 @@ namespace amanatApi.Models
         public string NO { get; set; }
         public string TITLE { get; set; }
         public int? SANADTYPE { get; set; }
-        public long SANADDATE { get; set; }
+        public int SANADDATE { get; set; }
         public int SANADCONDITION { get; set; }
         public double? TOTAL { get; set; }
         public int? PAPERID { get; set; }
-        public long CREATEDATE { get; set; }
+        public int CREATEDATE { get; set; }
         public int CREATEUSER { get; set; }
         public int? APPROVEDUSER { get; set; }
-        public long? APPROVEDDATE { get; set; }
+        public int? APPROVEDDATE { get; set; }
         public int YEAR { get; set; }
         public int? ORGANIZATION { get; set; }
         public int LOCATION { get; set; }
