+++ amanatApi/Models/ERP_BILL.cs
@@ -11,11 +11,11 @@ namespace amanatApi.Models
         public int ENSH_ID { get; set; }
         public int PARDAKHT_VAZIAT { get; set; }
         public int? PARDAKHAT_TYPE { get; set; }
-        public long? PARDAKHT_DATE { get; set; }
+        public int? PARDAKHT_DATE { get; set; }
         public double PRICE { get; set; }
         public string RAHGIRI_NO { get; set; }
-        public long BEGINDATE { get; set; }
-        public long ENDDATE { get; set; }
+        public int BEGINDATE { get; set; }
+        public int ENDDATE { get; set; }
         public string SHENASE_PARDAKHT { get; set; }
         public string SHENASE_GHABZ { get; set; }
         public int? APPROVER { get; set; }
@@ -38,7 +38,7 @@ namespace amanatApi.Models
         public double? BIME { get; set; }
         public double? AVAREZ { get; set; }
         public string XLSFILE { get; set; }
-        public long? CREATE_DATE { get; set; }
+        public int? CREATE_DATE { get; set; }
 
         public virtual ERP_ENSH ENSH { get; set; }
     }
