+++ amanatApi/Models/ERP_EXT384.cs
@@ -19,7 +19,7 @@ namespace amanatApi.Models
         public string PUK2 { get; set; }
         public string FILENO { get; set; }
         public string SABTNO { get; set; }
-        public long? SODORDATE { get; set; }
+        public int? SODORDATE { get; set; }
         public int? COUNTRYID { get; set; }
         public string STATE { get; set; }
         public string CITY { get; set; }
@@ -29,7 +29,7 @@ namespace amanatApi.Models
         public string DAFTARCODE { get; set; }
         public string DAFTERSABTNO { get; set; }
         public string DAFTARSAFHE { get; set; }
-        public long? DAFTARDATE { get; set; }
+        public int? DAFTARDATE { get; set; }
         public int? SANADTEDAD { get; set; }
         public int? SANADVAZIATID { get; set; }
         public int? SANADTYPEID { get; set; }
@@ -38,13 +38,13 @@ namespace amanatApi.Models
         public string FISHNO { get; set; }
         public string BANK { get; set; }
         public int? FISHVALUE { get; set; }
-        public long? FISHPARDAKHTDATE { get; set; }
+        public int? FISHPARDAKHTDATE { get; set; }
         public int? VEKALATNAMEID { get; set; }
         public int? EGHRARNAMEID { get; set; }
         public int? SHENASNAMEID { get; set; }
         public int? FISHTASVIRID { get; set; }
         public string SOLHNO { get; set; }
-        public long? SOLHDATE { get; set; }
+        public int? SOLHDATE { get; set; }
         public int? MALEKPID { get; set; }
     }
 }
