+++ amanatApi/Models/ERP_FIRMMEMBER.cs
@@ -12,15 +12,15 @@ namespace amanatApi.Models
         public string NAME { get; set; }
         public string LNAME { get; set; }
         public string FATHERNAME { get; set; }
-        public long? BIRTHDATE { get; set; }
+        public int? BIRTHDATE { get; set; }
         public int? NID { get; set; }
         public string TEL { get; set; }
         public string MOBILE { get; set; }
         public int? SEMATID { get; set; }
         public double? SHAREAMOUNT { get; set; }
         public int? SIGNRIGHT { get; set; }
-        public long? SIGNRIGHT_BEGIN { get; set; }
-        public long? SIGNRIGHT_END { get; set; }
+        public int? SIGNRIGHT_BEGIN { get; set; }
+        public int? SIGNRIGHT_END { get; set; }
         public int? STATUS { get; set; }
         public string ADDRESS { get; set; }
         public int? FAKEID { get; set; }
