+++ amanatApi/Models/ERP_MSGDETAIL.cs
@@ -8,9 +8,9 @@ namespace amanatApi.Models
     public partial class ERP_MSGDETAIL
     {
         public int VISITED { get; set; }
-        public long? DATE_VISIT { get; set; }
+        public int? DATE_VISIT { get; set; }
         public int REPLIED { get; set; }
-        public long? DATE_REPLY { get; set; }
+        public int? DATE_REPLY { get; set; }
         public int FORWARDED { get; set; }
         public int ARCHIVED { get; set; }
         public int ID { get; set; }
